% Generated by build script
\name{uniprot_search}
\alias{uniprot_search}
\title{Search UniProt proteins}
\usage{
uniprot_search()
}
\arguments{
\item{query}{Search term (gene name, protein name, etc.)}
\item{size}{Number of results (default 25)}
}
\value{tibble of matching proteins}
\description{Search UniProt proteins}
